Use u128 in progressive balance cache
Fixes PulseChain overflow condition introduced in upstream v4.3.0
release.

+++ consensus/fork_choice/src/fork_choice.rs
@@ -1610,9 +1610,9 @@ where
 fn check_progressive_balances<E, T>(
     state: &BeaconState<E>,
     participation_cache: &ParticipationCache,
-    cached_previous_target_balance: u64,
-    cached_current_target_balance: u64,
-    cached_total_active_balance: u64,
+    cached_previous_target_balance: u128,
+    cached_current_target_balance: u128,
+    cached_total_active_balance: u128,
 ) -> Result<(), Error<T::Error>>
 where
     E: EthSpec,
